+++ src/label_frame.rs
@@ -32,12 +32,12 @@ impl grid::TkGridLayout for TkLabelFrame {}
 
 impl TkLabelFrame {
     /// Size of border around frame
-    pub fn border_width(&self, width: u32) {
+    pub fn border_width(&self, width: u64) {
         widget::configure(&self.id, "borderwidth", &width.to_string());
     }
 
     /// Height of frame, in rows
-    pub fn height(&self, height: u32) {
+    pub fn height(&self, height: u64) {
         widget::configure(&self.id, "height", &height.to_string());
     }
 
@@ -70,7 +70,7 @@ impl TkLabelFrame {
     /// * [left-right top-bottom]
     /// * [left top-bottom right]
     /// * [left top right bottom]
-    pub fn padding(&self, values: &[u32]) {
+    pub fn padding(&self, values: &[u64]) {
         widget::padding(&self.id, values);
     }
 
@@ -90,12 +90,12 @@ impl TkLabelFrame {
     }
 
     /// Underlines the character at the given index position.
-    pub fn underline(&self, index: u32) {
+    pub fn underline(&self, index: u64) {
         widget::configure(&self.id, "underline", &index.to_string());
     }
 
     /// Width of frame, in columns
-    pub fn width(&self, width: u32) {
+    pub fn width(&self, width: u64) {
         widget::configure(&self.id, "width", &width.to_string());
     }
 }
