+++ src/scale.rs
@@ -34,7 +34,7 @@ impl grid::TkGridLayout for TkScale {}
 
 impl TkScale {
     /// Sets the function to be called whenever the scale value is changed.
-    pub fn command(&self, command: impl Fn(f32) + Send + 'static) {
+    pub fn command(&self, command: impl Fn(f64) + Send + 'static) {
         wish::add_callback1_float(&self.id, wish::mk_callback1_float(command));
         let msg = format!(
             "{} configure -command [list scale_value {}]",
@@ -44,20 +44,20 @@ impl TkScale {
     }
 
     /// Sets the minimum value for the scale.
-    pub fn from(&self, value: f32) {
+    pub fn from(&self, value: f64) {
         widget::configure(&self.id, "from", &value.to_string());
     }
 
     /// Sets the maximum value for the scale.
-    pub fn to(&self, value: f32) {
+    pub fn to(&self, value: f64) {
         widget::configure(&self.id, "to", &value.to_string());
     }
 
     /// Retrieves the scale's value.
-    pub fn value_get(&self) -> f32 {
+    pub fn value_get(&self) -> f64 {
         let msg = format!("puts [{} get] ; flush stdout", self.id);
         let result = wish::eval_wish(&msg);
-        if let Ok(value) = result.parse::<f32>() {
+        if let Ok(value) = result.parse::<f64>() {
             value
         } else {
             0.0
@@ -65,12 +65,12 @@ impl TkScale {
     }
 
     /// Set the scale's value.
-    pub fn value(&self, value: f32) {
+    pub fn value(&self, value: f64) {
         widget::configure(&self.id, "value", &value.to_string());
     }
 
     /// Displayed length of scale in pixels.
-    pub fn length(&self, value: u32) {
+    pub fn length(&self, value: u64) {
         widget::configure(&self.id, "length", &value.to_string());
     }
